+++ src/Twilio/Rest/Verify/V2/ServiceOptions.php
@@ -30,7 +30,8 @@ abstract class ServiceOptions {
      * @param bool $customCodeEnabled Whether to allow sending verifications with a
      *                                custom code.
      * @param bool $pushIncludeDate Optional. Include the date in the Challenge's
-     *                              reponse. Default: true
+     *                              response. Default: false. **Deprecated** do not
+     *                              use this parameter.
      * @param string $pushApnCredentialSid Optional. Set APN Credential for this
      *                                     service.
      * @param string $pushFcmCredentialSid Optional. Set FCM Credential for this
@@ -67,7 +68,8 @@ abstract class ServiceOptions {
      * @param bool $customCodeEnabled Whether to allow sending verifications with a
      *                                custom code.
      * @param bool $pushIncludeDate Optional. Include the date in the Challenge's
-     *                              reponse. Default: true
+     *                              response. Default: false. **Deprecated** do not
+     *                              use this parameter.
      * @param string $pushApnCredentialSid Optional. Set APN Credential for this
      *                                     service.
      * @param string $pushFcmCredentialSid Optional. Set FCM Credential for this
@@ -105,7 +107,8 @@ class CreateServiceOptions extends Options {
      * @param bool $customCodeEnabled Whether to allow sending verifications with a
      *                                custom code.
      * @param bool $pushIncludeDate Optional. Include the date in the Challenge's
-     *                              reponse. Default: true
+     *                              response. Default: false. **Deprecated** do not
+     *                              use this parameter.
      * @param string $pushApnCredentialSid Optional. Set APN Credential for this
      *                                     service.
      * @param string $pushFcmCredentialSid Optional. Set FCM Credential for this
@@ -234,10 +237,11 @@ class CreateServiceOptions extends Options {
     }
 
     /**
-     * Optional configuration for the Push factors. If true, include the date in the Challenge's reponse. Otherwise, the date is omitted from the response. See [Challenge](https://www.twilio.com/docs/verify/api/challenge) resource’s details parameter for more info. Default: true
+     * Optional configuration for the Push factors. If true, include the date in the Challenge's response. Otherwise, the date is omitted from the response. See [Challenge](https://www.twilio.com/docs/verify/api/challenge) resource’s details parameter for more info. Default: false. **Deprecated** do not use this parameter. This timestamp value is the same one as the one found in `date_created`, please use that one instead.
      *
      * @param bool $pushIncludeDate Optional. Include the date in the Challenge's
-     *                              reponse. Default: true
+     *                              response. Default: false. **Deprecated** do not
+     *                              use this parameter.
      * @return $this Fluent Builder
      */
     public function setPushIncludeDate(bool $pushIncludeDate): self {
@@ -357,7 +361,8 @@ class UpdateServiceOptions extends Options {
      * @param bool $customCodeEnabled Whether to allow sending verifications with a
      *                                custom code.
      * @param bool $pushIncludeDate Optional. Include the date in the Challenge's
-     *                              reponse. Default: true
+     *                              response. Default: false. **Deprecated** do not
+     *                              use this parameter.
      * @param string $pushApnCredentialSid Optional. Set APN Credential for this
      *                                     service.
      * @param string $pushFcmCredentialSid Optional. Set FCM Credential for this
@@ -498,10 +503,11 @@ class UpdateServiceOptions extends Options {
     }
 
     /**
-     * Optional configuration for the Push factors. If true, include the date in the Challenge's reponse. Otherwise, the date is omitted from the response. See [Challenge](https://www.twilio.com/docs/verify/api/challenge) resource’s details parameter for more info. Default: true
+     * Optional configuration for the Push factors. If true, include the date in the Challenge's response. Otherwise, the date is omitted from the response. See [Challenge](https://www.twilio.com/docs/verify/api/challenge) resource’s details parameter for more info. Default: false. **Deprecated** do not use this parameter.
      *
      * @param bool $pushIncludeDate Optional. Include the date in the Challenge's
-     *                              reponse. Default: true
+     *                              response. Default: false. **Deprecated** do not
+     *                              use this parameter.
      * @return $this Fluent Builder
      */
     public function setPushIncludeDate(bool $pushIncludeDate): self {