+++ src/Twilio/Rest/Video/V1/RoomList.php
@@ -57,6 +57,7 @@ class RoomList extends ListResource {
             'MaxParticipantDuration' => $options['maxParticipantDuration'],
             'EmptyRoomTimeout' => $options['emptyRoomTimeout'],
             'UnusedRoomTimeout' => $options['unusedRoomTimeout'],
+            'LargeRoom' => Serialize::booleanToString($options['largeRoom']),
         ]);
 
         $payload = $this->version->create('POST', $this->uri, [], $data);