+++ tests/Twilio/Integration/Verify/V2/ServiceTest.php
@@ -52,7 +52,7 @@ class ServiceTest extends HolodeckTestCase {
                 "do_not_share_warning_enabled": false,
                 "custom_code_enabled": true,
                 "push": {
-                    "include_date": true,
+                    "include_date": false,
                     "apn_credential_sid": "CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                     "fcm_credential_sid": null
                 },
@@ -116,7 +116,7 @@ class ServiceTest extends HolodeckTestCase {
                 "do_not_share_warning_enabled": false,
                 "custom_code_enabled": true,
                 "push": {
-                    "include_date": true,
+                    "include_date": false,
                     "apn_credential_sid": null,
                     "fcm_credential_sid": null
                 },
@@ -216,7 +216,7 @@ class ServiceTest extends HolodeckTestCase {
                         "do_not_share_warning_enabled": false,
                         "custom_code_enabled": true,
                         "push": {
-                            "include_date": true,
+                            "include_date": false,
                             "apn_credential_sid": null,
                             "fcm_credential_sid": null
                         },
@@ -282,7 +282,7 @@ class ServiceTest extends HolodeckTestCase {
                 "do_not_share_warning_enabled": false,
                 "custom_code_enabled": true,
                 "push": {
-                    "include_date": true,
+                    "include_date": false,
                     "apn_credential_sid": null,
                     "fcm_credential_sid": "CRbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
                 },