+++ tests/Twilio/Integration/Video/V1/RoomTest.php
@@ -58,6 +58,7 @@ class RoomTest extends HolodeckTestCase {
                 "empty_room_timeout": 5,
                 "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
+                "large_room": false,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -115,6 +116,7 @@ class RoomTest extends HolodeckTestCase {
                 "empty_room_timeout": 5,
                 "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
+                "large_room": false,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -158,6 +160,7 @@ class RoomTest extends HolodeckTestCase {
                 "empty_room_timeout": 5,
                 "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
+                "large_room": false,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -201,6 +204,7 @@ class RoomTest extends HolodeckTestCase {
                 "empty_room_timeout": 5,
                 "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
+                "large_room": false,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -242,6 +246,7 @@ class RoomTest extends HolodeckTestCase {
                 "empty_room_timeout": 5,
                 "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
+                "large_room": false,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -285,6 +290,7 @@ class RoomTest extends HolodeckTestCase {
                 "empty_room_timeout": 5,
                 "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
+                "large_room": false,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -328,6 +334,7 @@ class RoomTest extends HolodeckTestCase {
                 "empty_room_timeout": 5,
                 "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
+                "large_room": false,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -369,6 +376,7 @@ class RoomTest extends HolodeckTestCase {
                 "empty_room_timeout": 5,
                 "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
+                "large_room": false,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -452,6 +460,7 @@ class RoomTest extends HolodeckTestCase {
                         "audio_only": false,
                         "empty_room_timeout": 5,
                         "unused_room_timeout": 5,
+                        "large_room": false,
                         "url": "https://video.twilio.com/v1/Rooms/RM4070b618362c1682b2385b1f9982833c",
                         "links": {
                             "participants": "https://video.twilio.com/v1/Rooms/RM4070b618362c1682b2385b1f9982833c/Participants",
@@ -523,6 +532,7 @@ class RoomTest extends HolodeckTestCase {
                 "empty_room_timeout": 5,
                 "unused_room_timeout": 5,
                 "end_time": "2015-07-30T20:00:00Z",
+                "large_room": false,
                 "duration": 10,
                 "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {